astar: use float consistently for PSVECDotProduct/PSVECMag results
- Remove unnecessary double casts in getEscapePos function
- PSVECDotProduct and PSVECMag return float; use float throughout
- Improves type consistency for PowerPC floating-point operations

+++ src/astar.cpp
@@ -736,10 +736,8 @@ CAStar::CAPos* CAStar::getEscapePos(Vec& from, Vec& base, int startGroup, int fo
 
 		dirToPortal.Normalize();
 
-		double dot = static_cast<double>(
-			PSVECDotProduct(reinterpret_cast<Vec*>(&escapeDir),
-			                reinterpret_cast<Vec*>(&dirToPortal))
-		);
+		float dot = PSVECDotProduct(reinterpret_cast<Vec*>(&escapeDir),
+		                            reinterpret_cast<Vec*>(&dirToPortal));
 
 		CVector baseVec2(base);
 		CVector portalPos2(portal.m_position);
@@ -756,9 +754,7 @@ CAStar::CAPos* CAStar::getEscapePos(Vec& from, Vec& base, int startGroup, int fo
 		diffForMag.y = diffBaseToPortalMagSrc.y;
 		diffForMag.z = diffBaseToPortalMagSrc.z;
 
-		double dist = static_cast<double>(
-			PSVECMag(reinterpret_cast<Vec*>(&diffForMag))
-		);
+		float dist = PSVECMag(reinterpret_cast<Vec*>(&diffForMag));
 
 		if (dot < 0.0)
 		{
